pyproject.toml
[project]
name = "langchain-starter-kit"
version = "0.2.0"
description = ""
authors = [
    {name = "jalakoo",email = "jalakoo@gmail.com"}
]
license = {text = "Apache 2.0"}
readme = "README.md"
requires-python = ">=3.12,<4.0"
dependencies = [
    "langchain (>=0.3.26,<0.4.0)",
    "langchain-neo4j (>=0.4.0,<0.5.0)",
    "langchain-openai (>=0.3.28,<0.4.0)",
    "fastapi (>=0.116.1,<0.117.0)",
    "uvicorn (>=0.35.0,<0.36.0)",
    "neo4j (>=5.28.1,<6.0.0)",
    "langchain-community (>=0.3.27,<0.4.0)",
    "python-dotenv (>=1.1.1,<2.0.0)"
]

[build-system]
requires = ["poetry-core>=2.0.0,<3.0.0"]
build-backend = "poetry.core.masonry.api"
